+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/theme/template.kt
@@ -110,7 +110,8 @@ object Template {
 
     var basketPreviewHeaderTemplate: (@Composable() (
         recipeVm: RecipeViewModel,
-        goBackToRecipeDetails: () -> Unit
+        goBackToRecipeDetails: () -> Unit,
+        previous: () -> Unit
     ) -> Unit)? = null