+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/basketPreview/BasketPreview.kt
@@ -30,6 +30,7 @@ class BasketPreview(
     recipeId: String,
     private val recipeVm: RecipeViewModel,
     val goToDetail: () -> Unit,
+    val previous: () -> Unit,
     val close: () -> Unit,
     private val goToItemSelector: () -> Unit,
 ) {
@@ -39,13 +40,8 @@ class BasketPreview(
     @ExperimentalCoilApi
     @Composable
     fun content() {
-
         fun removeRecipeAndClose() {
-            vmBasketPreview.setEvent(
-                BasketPreviewContract.Event.RemoveRecipe(
-                    recipeVm.recipeId ?: ""
-                )
-            )
+            vmBasketPreview.setEvent(BasketPreviewContract.Event.RemoveRecipe(recipeVm.recipeId ?: ""))
             close()
         }
 
@@ -53,12 +49,10 @@ class BasketPreview(
             topBar = {
                 if (Template.basketPreviewHeaderTemplate != null) {
                     Template.basketPreviewHeaderTemplate?.let {
-                        it(recipeVm) { goToDetail() }
+                        it(recipeVm, goToDetail, previous)
                     }
                 } else {
-                    BasketPreviewHeader {
-                        goToDetail()
-                    }
+                    BasketPreviewHeader(goToDetail, previous)
                 }
             },
             content = {