+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/catalog/Catalog.kt
@@ -55,8 +55,6 @@ class Catalog @JvmOverloads constructor(
     override fun Content() {
 
         val state by vmCatalog.uiState.collectAsState()
-
-
         Box {
             Column {
                 ManagementResourceState(