+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/catalog/CatalogSearch.kt
@@ -24,6 +24,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.focus.FocusRequester
@@ -32,34 +33,44 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
 import com.miam.kmmMiamCore.component.singletonFilter.SingletonFilterContract
 import com.miam.kmmMiamCore.component.singletonFilter.SingletonFilterViewModel
+import com.miam.kmmMiamCore.services.RouteService
 import com.miam.kmm_miam_sdk.android.theme.Colors
 import com.miam.kmm_miam_sdk.android.theme.Colors.white
 import com.miam.kmm_miam_sdk.android.theme.Template
 import com.miam.kmm_miam_sdk.android.ui.components.catalog.customization.CatalogImage
 import com.miam.kmm_miam_sdk.android.ui.components.catalog.customization.CatalogImage.search
 import com.miam.kmm_miam_sdk.android.ui.components.common.Clickable
-import com.miam.kmm_miam_sdk.android.ui.components.routerOutlet.FullScreen
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
 
 // TODO Refact with filter service
 class CatalogSearch(
     private val catalogFilterVM: SingletonFilterViewModel,
     private val closeDialog: () -> Unit,
     private val goToSearchResult: () -> Unit,
-) {
+): KoinComponent {
 
     private fun updateResearch(searchString: String) {
         catalogFilterVM.setSearchString(searchString)
     }
 
+    private val routeService: RouteService by inject()
+
+    @OptIn(ExperimentalComposeUiApi::class)
     @Composable
     fun Content() {
 
         val state = catalogFilterVM.uiState.collectAsState()
         val textState = remember { mutableStateOf(state.value.searchString ?: "") }
 
-        FullScreen {
+        Dialog(
+            onDismissRequest = { routeService.previous() },
+            properties = DialogProperties(usePlatformDefaultWidth = false)
+        ) {
             if (Template.CatalogSearchTemplate != null) {
                 TemplateView(state)
             } else {