+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/router/RouterOutletContract.kt
@@ -12,20 +12,13 @@ enum class RouterContent {
 interface RouterOutletContract {
 
     sealed class Event: UiEvent {
-        object GoToHelper: RouterOutletContract.Event()
-        object GoToSponsor: RouterOutletContract.Event()
-        object OpenDialog: RouterOutletContract.Event()
-        object CloseDialog: RouterOutletContract.Event()
-        object CloseDialogFromPreview: RouterOutletContract.Event()
-        object GoToItemSelector: RouterOutletContract.Event()
-        data class GoToPreview(val recipeId: String, val vm: RecipeViewModel):
-            RouterOutletContract.Event()
-
-        data class GoToDetail(val vm: RecipeViewModel, val withFooter: Boolean = true):
-            RouterOutletContract.Event()
-
-        data class GoToDetailFromPreview(val vm: RecipeViewModel): RouterOutletContract.Event()
-        data class SetRouterContent(val routerContent: RouterContent): RouterOutletContract.Event()
+        object GoToHelper: Event()
+        object GoToSponsor: Event()
+        object OpenDialog: Event()
+        object GoToItemSelector: Event()
+        data class GoToPreview(val recipeId: String, val vm: RecipeViewModel): Event()
+        data class GoToDetail(val vm: RecipeViewModel, val withFooter: Boolean = true): Event()
+        object Previous: Event()
     }
 
     data class State(