+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/router/RouterOutletViewModel.kt
@@ -4,11 +4,14 @@ import com.miam.kmmMiamCore.base.mvi.BaseViewModel
 import com.miam.kmmMiamCore.component.recipe.RecipeViewModel
 import com.miam.kmmMiamCore.handler.LogHandler
 import com.miam.kmmMiamCore.services.Analytics
+import com.miam.kmmMiamCore.services.RouteService
+import com.miam.kmmMiamCore.services.RouteServiceAction
 import org.koin.core.component.inject
 
 open class RouterOutletViewModel: BaseViewModel<RouterOutletContract.Event, RouterOutletContract.State, RouterOutletContract.Effect>() {
 
     private val analyticsService: Analytics by inject()
+    private val routeService: RouteService by inject()
 
     fun goToDetail(vmRecipe: RecipeViewModel, showDetailsFooter: Boolean = true) {
         LogHandler.info("Miam RouterOutletViewModel goToDetail $vmRecipe")
@@ -29,64 +32,33 @@ open class RouterOutletViewModel: BaseViewModel<RouterOutletContract.Event, Rout
         )
     }
 
-
     override fun handleEvent(event: RouterOutletContract.Event) {
         when (event) {
             is RouterOutletContract.Event.GoToDetail -> {
                 LogHandler.info("Miam RouterOutletViewModel goToDetail event $event")
                 // TODO : path
-                analyticsService.sendEvent(
-                    Analytics.EVENT_PAGEVIEW,
-                    "/detail",
-                    Analytics.PlausibleProps(recipe_id = event.vm.recipeId)
-                )
-                analyticsService.sendEvent(
-                    Analytics.EVENT_RECIPE_DISPLAY,
-                    "",
-                    Analytics.PlausibleProps(recipe_id = event.vm.recipeId)
-                )
+                analyticsService.sendEvent(Analytics.EVENT_PAGEVIEW, "/detail", Analytics.PlausibleProps(recipe_id = event.vm.recipeId))
+                analyticsService.sendEvent(Analytics.EVENT_RECIPE_DISPLAY, "", Analytics.PlausibleProps(recipe_id = event.vm.recipeId))
                 setState { copy(rvm = event.vm, showFooter = event.withFooter) }
                 LogHandler.info("Miam RouterOutletViewModel will navigate")
+                routeService.dispatch(RouteServiceAction.SetDialogRoute("Detail", { navigateTo(RouterContent.RECIPE_DETAIL) }, ::onClose))
                 navigateTo(RouterContent.RECIPE_DETAIL)
             }
             is RouterOutletContract.Event.GoToPreview -> {
                 LogHandler.info("Miam RouterOutletViewModel GoToPreview event $event")
                 // TODO : path
-                analyticsService.sendEvent(
-                    Analytics.EVENT_PAGEVIEW,
-                    "/basket-preview",
-                    Analytics.PlausibleProps(recipe_id = event.recipeId)
-                )
-                setState {
-                    copy(
-                        recipeId = event.recipeId,
-                        rvm = event.vm
-                    )
-                }
+                analyticsService.sendEvent(Analytics.EVENT_PAGEVIEW, "/basket-preview", Analytics.PlausibleProps(recipe_id = event.recipeId))
+                setState { copy(recipeId = event.recipeId, rvm = event.vm) }
+                routeService.dispatch(RouteServiceAction.SetDialogRoute("Preview", { navigateTo(RouterContent.BASKET_PREVIEW) }, ::onClose))
                 navigateTo(RouterContent.BASKET_PREVIEW)
             }
             is RouterOutletContract.Event.GoToItemSelector -> {
                 // TODO : path
-                analyticsService.sendEvent(
-                    Analytics.EVENT_PAGEVIEW,
-                    "/replace-item",
-                    Analytics.PlausibleProps(recipe_id = currentState.recipeId)
-                )
+                analyticsService.sendEvent(Analytics.EVENT_PAGEVIEW, "/replace-item", Analytics.PlausibleProps(recipe_id = currentState.recipeId))
+                routeService.dispatch(RouteServiceAction.SetDialogRoute("Item Selector", { navigateTo(RouterContent.ITEMS_SELECTOR) }, ::onClose))
                 navigateTo(RouterContent.ITEMS_SELECTOR)
             }
-            is RouterOutletContract.Event.CloseDialogFromPreview -> {
-                setEvent(RouterOutletContract.Event.CloseDialog)
-            }
-            is RouterOutletContract.Event.GoToDetailFromPreview -> {
-                setEvent(RouterOutletContract.Event.GoToDetail(event.vm))
-            }
-            is RouterOutletContract.Event.SetRouterContent -> {
-                setState {
-                    copy(
-                        content = event.routerContent
-                    )
-                }
-            }
+            is RouterOutletContract.Event.Previous -> routeService.previous()
             RouterOutletContract.Event.GoToHelper -> navigateTo(RouterContent.RECIPE_HELPER)
             RouterOutletContract.Event.GoToSponsor -> navigateTo(RouterContent.RECIPE_SPONSOR)
             RouterOutletContract.Event.OpenDialog -> {
@@ -94,8 +66,11 @@ open class RouterOutletViewModel: BaseViewModel<RouterOutletContract.Event, Rout
                 setState { copy(isOpen = true) }
                 LogHandler.info("Miam RouterOutletViewModel will OpenDialog ${this.currentState}")
             }
-            RouterOutletContract.Event.CloseDialog -> setState { copy(isOpen = false) }
         }
+    }
+
+    private fun onClose() {
+        setState { copy(isOpen = false) }
     }
 
     private fun navigateTo(destination: RouterContent) {